requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
lxml==4.4.0
numpy==1.17.0
Pillow==6.1.0
pyspellchecker==0.5.0
requests==2.22.0
urllib3==1.25.3
